動作確認テストについてvitestだとdetectopenhandlesが無いので一つの単純なtypescriptの実行ファイルにしてtsxで実行するように変更してください
